Boolkaco/#56/enable iban selector fiat routes (#69)
* #56 - DFX Integration part 1 - add list unique ibans

* #56 - DFX integration part 1 - general buy/sell routes

+++ lib/bloc/fiat/fiat_state.dart
@@ -19,7 +19,7 @@ class FiatState extends Equatable {
   final List<dynamic>? countryList;
   final List<FiatModel>? fiatList;
   final bool? isShowTutorial;
-  final double? limit;
+  final int? limit;
 
   FiatState({
     this.status = FiatStatusList.initial,
@@ -79,7 +79,7 @@ class FiatState extends Equatable {
     List<dynamic>? countryList,
     List<FiatModel>? fiatList,
     bool? isShowTutorial,
-    double? limit,
+    int? limit,
   }) {
     return FiatState(
       status: status ?? this.status,
